requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Flask>=3.0.0
torch
torchvision
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
gunicorn>=21.2.0
timm
